Only clear /dist on prod build
This is running everytime and sometimes I want to look at the /dist without having to rebuild =P

+++ src/builder/index.js
@@ -35,7 +35,12 @@ export default function(config: Object): void {
   config.webpackConfigNode = webpackConfigNode(config)
 
   const destination = join(config.cwd, config.destination)
-  fs.emptyDirSync(destination)
+  
+  if (config.production) {
+    // If production build, clear /dist folder
+    fs.emptyDirSync(destination)
+  }
+
 
   if (config.static) {
     // Copy static assets to build folder
